requirements.txt
-sqlite3==0.0.0
Django==1.6
MySQL-python==1.2.5
Pillow==2.4.0
South==0.8.4
django-allauth==0.17.0
django-braces==1.4.0
django-guardian==1.2.0
django-summernote==0.5.12
django-widget-tweaks==1.3
meld3==0.6.8
oauthlib==0.6.3
python-openid==2.2.5
pytils==0.3
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.2
stevedore==0.15
supervisor==3.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.2
wsgiref==0.1.2
